api_chain/requirements.txt
langchain==0.2.6
langchain-community==0.2.5
langchain-core==0.2.10
langchain-google-genai
langchain-huggingface
langchain-chroma
ipykernel
ipywidgets
widgetsnbextension
pandas-profiling
jupyter
hvac
xformers
langgraph
surrealdb
torch==2.4.0+cu124
torchvision==0.15.1+cu124
torchaudio==0.10.1+cu124
python-dotenv
-f https://download.pytorch.org/whl/cu124
